refactor metadata libraries
- create rdiffbackup.meta namespace
- split metadata into meta_mgr, meta and meta.attr for different purposes
- split eas_acls into meta.ea and meta.acl_posix
- move win_acls to meta.acl_win
- keep stubs for backward compatibility of eas_acls and win_acls

DEV: improved execution output during tests so that commands can be more easily reused
DEV: align more actions and meta plugins interfaces, add generic plugins module
DEV: improve plugins documentation and add meta plugins docs
