+++ docs/arch/plugins/actions.adoc
@@ -1,4 +1,4 @@
-= RDIFF-BACKUP ACTIONS PLUGINS
+= Actions plugins
 :sectnums:
 :toc:
 
@@ -9,10 +9,10 @@ An action plug-in is:
 . either in the `rdiffbackup.actions` namespace, with any module name (this is how built-in action modules are packaged),
 . or a top-level module with a name `+rdb_action_[...].py+`
 
-All those plug-ins are found by the action manager module `rdiffbackup.actions_mgr` and returned as a hash made of key-value pairs by the function `get_discovered_actions`:
+All those plug-ins are found by the action manager module `rdiffbackup.actions_mgr` and returned as a hash made of key-value pairs by the function `get_actions_dict`:
 
-* key is the name returned by the Action class (see below the `get_name` method)
-* value is the Action class itself, as returned by the `get_action_class` function of the action plug-in module.
+* key is the name returned by the Action class
+* value is the Action class itself, as returned by the `get_plugin_class` function of the action plug-in module.
 
 NOTE: more information is provided with `pydoc rdiffbackup.actions_mgr`.
 
@@ -21,21 +21,18 @@ The actions manager also has two functions to get lists of argparse parsers to u
 == Module interface
 
 Each action plug-in module contains one single Action-class.
-This action class is returned by the function `get_action_class` of the module.
+This action class is returned by the function `get_plugin_class` of the module.
 How the action class is named and of which hierarchy it descends is irrelevant (though it makes sense to derive it from the rdiffbackup.actions.BaseAction class).
 
 All further action plug-in interactions are executed through the interface of this single class.
 
 == Action class interface
 
-The Action class has the following interface:
+The Action class has the following interface, in addition to the generic plugin class requirements:
 
-* the class method `cls.get_name()` returns the name of the plug-in, only this name will be used in the code, and it doesn't need to be aligned with the name of the module or of the class (but it should).
-Only requirement is that it is unique or plug-ins will overwrite each other.
 * the class method `cls.get_security_class()` returns the security class of the action plug-in, one of `backup`, `restore`, `server` or `validate`.
 A security class of `None` is only applicable to actions without client/server context (e.g.
 `info`).
-* the class method `cls.get_version()` returns the version of the plug-in as a string.
 * the class method `cls.add_action_subparser(sub_handler)` returns a subparser as returned by argparse's `sub_handler.add_parser` function, so that the sub-options of the action can be parsed by the `rdiffbackup.arguments.parse` function.
 * the method `self.__init__(args, log, errlog)` initializes the class as object, based on the namespace returned by argparse, a Log and an ErrorLog object.
 * the method `self.pre_check()` just validates that the arguments passed were correct, beyond what argparse could do, and shouldn't try to connect yet to any remote location.
@@ -68,7 +65,7 @@ NOTE: more information is provided with `pydoc rdiffbackup.actions`.
 Taking all together, the code to use an action plugin would look as follows, without any error handling:
 
 ----
-discovered_actions = actions_mgr.get_discovered_actions()
+discovered_actions = actions_mgr.get_actions_dict()
 parsed_args = arguments.parse(
 	arglist, "rdiff-backup {ver}".format(ver=Globals.version),
 	actions_mgr.get_generic_parsers(),
