+++ src/rdiffbackup/actions/calculate.py
@@ -60,5 +60,5 @@ class CalculateAction(actions.BaseAction):
         return 0
 
 
-def get_action_class():
+def get_plugin_class():
     return CalculateAction
