+++ src/rdiffbackup/actions/server.py
@@ -90,5 +90,5 @@ class ServerAction(actions.BaseAction):
                     log.Log.WARNING)
 
 
-def get_action_class():
+def get_plugin_class():
     return ServerAction
