+++ testing/hashtest.py
@@ -6,7 +6,8 @@ import sys
 import os
 from commontest import abs_test_dir, re_init_rpath_dir, Myrm, \
     abs_output_dir, rdiff_backup, abs_testing_dir, MakeOutputDir
-from rdiff_backup import hash, rpath, metadata, Globals, Security, SetConnections
+from rdiff_backup import hash, rpath, Globals, Security, SetConnections
+from rdiffbackup.meta import attr
 
 
 class HashTest(unittest.TestCase):
@@ -73,7 +74,7 @@ class HashTest(unittest.TestCase):
         """Return list of hashes in the metadata_rp"""
         result = []
         comp = metadata_rp.isinccompressed()
-        extractor = metadata.RorpExtractor(metadata_rp.open("r", comp))
+        extractor = attr.AttrExtractor(metadata_rp.open("r", comp))
         for rorp in extractor.iterate():
             if rorp.has_sha1():
                 result.append(rorp.get_sha1())
@@ -81,8 +82,8 @@ class HashTest(unittest.TestCase):
                 result.append(None)
         return result
 
-    @unittest.skip("Skipping until hash of hard links is fixed, see issue #23."
-                   )
+    @unittest.skip(
+        "Skipping until hash of hard links is fixed, see issue #23.")
     def test_session(self):
         """Run actual sessions and make sure proper hashes recorded
